Added New Elements
Added a Salary field for keeping track of job salaries, if not in the table, prints N/A.

Added in error handling for certain text, so database is now UTF8mb4

Added sticking process button due to massive amount of job application backup in the database, might change to pagination in the future not sure.

Changed from alerts to using bootstraps "toast" css styling.

Changed order of baseline sort to have NEW jobs at the top.