Added Resume Upload and Parsing Functionality
These additions give the user the ability to upload a resume in a form, which then sends it to the backend, which runs a script that parses the resume, and currently returns the parsed data.

Eventually I will use NLP, or something along the sort to generate a score for each job based on a resume upload. Not sure the plan of action as of yet.